+++ src/lib.rs
@@ -41,4 +41,3 @@ pub mod tree;
 // pub mod symbols    L27
 // pub mod tree       L28
 // ----------------------
-